Move EditorContext to its own file with useEditor custom hook
