+++ frontend/src/components/UserNavigation.jsx
@@ -1,7 +1,7 @@
 import AnimationWrapper from "./AnimationWrapper";
 import { Link } from "react-router-dom";
 import { useContext } from "react";
-import { UserContext } from "../context/UserProvider";
+import { UserContext } from "../context/UserContext";
 import { removeFromSession } from "../services/session";
 
 const UserNavigation = () => {
@@ -21,32 +21,20 @@ const UserNavigation = () => {
 			transition={{ duration: 0.2 }}
 		>
 			<div className="bg-white absolute right-0 border border-grey w-60 duration-200">
-				<Link
-					to="/editor"
-					className="flex gap-2 link md:hidden pl-8 py-4"
-				>
+				<Link to="/editor" className="flex gap-2 link md:hidden pl-8 py-4">
 					<i className="fi fi-rr-file-edit"></i>
 					<p>Write</p>
 				</Link>
 
-				<Link
-					to={`/user/${username}`}
-					className="link pl-8 py-4"
-				>
+				<Link to={`/user/${username}`} className="link pl-8 py-4">
 					Profile
 				</Link>
 
-				<Link
-					to="/dashboard/blogs"
-					className="link pl-8 py-4"
-				>
+				<Link to="/dashboard/blogs" className="link pl-8 py-4">
 					Dashboard
 				</Link>
 
-				<Link
-					to="/settings/edit-profile"
-					className="link pl-8 py-4"
-				>
+				<Link to="/settings/edit-profile" className="link pl-8 py-4">
 					Settings
 				</Link>
