+++ frontend/src/hooks/useEditor.js
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { EditorContext } from "../context/EditorContext";
+
+const useEditor = () => {
+	return useContext(EditorContext);
+};
+
+export default useEditor;
