+++ frontend/src/pages/Editor.jsx
@@ -1,51 +1,29 @@
-import { useState, useContext, createContext } from "react";
+import { useContext, createContext } from "react";
 import { Navigate } from "react-router-dom";
-import { UserContext } from "../context/UserProvider";
+import { UserContext } from "../context/UserContext";
+import { EditorProvider } from "../context/EditorContext";
+import useEditor from "../hooks/useEditor";
 import BlogEditor from "../components/BlogEditor";
 import PublishForm from "../components/PublishForm";
-
-// Define starting state of editor
-const blogStructure = {
-	title: "",
-	banner: "",
-	content: [],
-	tags: [],
-	description: "",
-	author: {
-		personal_info: {},
-	},
-};
 
 export const EditorContext = createContext({});
 
 const Editor = () => {
-	const [editorState, setEditorState] = useState("editor");
-	const [blog, setBlog] = useState(blogStructure);
-	const [textEditor, setTextEditor] = useState({ isReady: false });
-
 	const {
 		userAuth: { access_token },
 	} = useContext(UserContext);
 
 	return (
-		<EditorContext.Provider
-			value={{
-				blog,
-				setBlog,
-				editorState,
-				setEditorState,
-				textEditor,
-				setTextEditor,
-			}}
-		>
-			{access_token === null ? (
-				<Navigate to="/signin" />
-			) : editorState === "editor" ? (
-				<BlogEditor />
-			) : (
-				<PublishForm />
-			)}
-		</EditorContext.Provider>
+		<EditorProvider>
+			{access_token === null ? <Navigate to="/signin" /> : <EditorContent />}
+		</EditorProvider>
 	);
 };
+
+const EditorContent = () => {
+	const { editorState } = useEditor();
+
+	return editorState === "editor" ? <BlogEditor /> : <PublishForm />;
+};
+
 export default Editor;
